requirements-generate-data.txt
lorem==0.1.1
Jinja2==2.11.3
pydbgen==1.0.5
Faker==2.0.5
alive-progress==2.4.1
PyYAML==6.0
